+++ coding/rouge/lib/rouge/lexers/logtalk.rb
@@ -62,6 +62,7 @@ module Rouge
         # Control constructs
         rule /(call|catch|throw)(?=[(])/, Keyword
         rule /(fail|false|true)\b/, Keyword
+        rule /(instantiation_error|type_error|domain_error|existence_error|permission_error|representation_error|evaluation_error|resource_error)(?=[(])/, Keyword
         # All solutions
         rule /(bagof|setof|findall|forall)(?=[(])/, Keyword
         # Multi-threading meta-predicates