+++ RELEASE_NOTES.md
@@ -49,6 +49,11 @@ Tests
 
 * ADDED: Unit tests for the new error throwing built-in methods.
 
+IDEs, text editors, and syntax highlighters support
+---------------------------------------------------
+
+* ADDED: Syntax coloring support for the new error throwing built-in methods.
+
 
 3.13.0 - November 8, 2017
 =========================